feat: model features and Python parity
Brings C++ inference to parity with the reference Python implementation
and adds several model capabilities:

- Support Qwen3-TTS 1.7B models with MTP projection bridging the
  2048-dim talker and 1024-dim code predictor (dynamic model detection)
- Fix UTF-8 tokenization with proper GPT-2 regex pre-tokenization,
  vendoring tokenizer_unicode from llama.cpp
- Wire voice steering instructions as separate prefill embeddings
  (matches reference, exposed via --instructions and server API)
- Add ICL voice cloning via Mimi codec encoder: reference audio is
  encoded to discrete speech codes combined with ref_text in prefill
- Support WAVE_FORMAT_EXTENSIBLE WAV headers (macOS screen recordings)
- Fix vocoder codebook normalization to use ggml_backend_tensor_get/set
  (GPU-safe, unbreaks Vulkan backend)
- Add Q8_0 tokenizer auto-selection with F16 fallback

Squashes upstream work from endo5501 (WAVE_FORMAT_EXTENSIBLE, 1.7B MTP
support) and adapts perf/feature work from Danmoreng/qwen3-tts.cpp.

+++ src/gguf_loader.cpp
@@ -203,7 +203,7 @@ bool load_tensor_data_from_file(
         
         read_buf.resize(nbytes);
         
-        if (fseek(f, data_offset + offset, SEEK_SET) != 0) {
+        if (fseek(f, (long)(data_offset + offset), SEEK_SET) != 0) {
             error_msg = "Failed to seek to tensor data: " + std::string(name);
             fclose(f);
             ggml_backend_free(backend);
